feat(bento): main file is now index.ts
