+++ package.json
@@ -5,8 +5,8 @@
     "test": "is-ci \"test:coverage\" \"test:watch\"",
     "test:watch": "jest --watch",
     "test:coverage": "jest --watch=false --coverage --passWithNoTests --updateSnapshot",
-    "build": "export NODE_ENV='production' && babel src -d lib --extensions \".ts\" --extensions \".js\"",
-    "build:ts": "tsc",
+    "build:babel": "export NODE_ENV='production' && babel src -d lib --extensions \".ts\" --extensions \".js\"",
+    "build": "tsc",
     "build:watch": "export NODE_ENV='development' && nodemon --watch src/ --exec \"babel src -d dist\"",
     "storybook:genreq": "node stories/generateRequireStories",
     "start": "start-storybook -p 9009 --no-dll",
