+++ src/index.ts
@@ -6,7 +6,7 @@ import "regenerator-runtime/runtime.js"
 /**
  * Animations
  */
-export {
+import {
   AppearIn,
   BounceIn,
   BounceOut,
@@ -20,11 +20,10 @@ export {
   SlideInFromRight,
   SlideInFromTop
 } from "./animations"
-
 /**
  * Formatting
  */
-export {
+import {
   abbreviateCurrency,
   CURRENCY_DISPLAY,
   dashIt,
@@ -41,11 +40,10 @@ export {
   UNIT_DISPLAY,
   UNITS
 } from "./formatting"
-
 /**
  * Math functions
  */
-export {
+import {
   almostEquals,
   clamp,
   cosineInterpolate,
@@ -54,11 +52,219 @@ export {
   lerp,
   multiplyPixelValue
 } from "./math"
-
 /**
  * Open-Id
  */
-export {
+import {
+  CallbackLogin,
+  CallbackLoginSilent,
+  CallbackLogout,
+  Logout,
+  OpenId,
+  OpenIdContext,
+  useCallbackLogin,
+  useCallbackLoginSilent,
+  useCallbackLogout,
+  useLogout,
+  useOpenId,
+  useOpenIdUser,
+  useOtpLogin,
+  useUrlLogin
+} from "./open-id"
+/**
+ * Sanity specific
+ */
+import {
+  BlockContent,
+  Document,
+  Image as SanityImage,
+  List as SanityList,
+  Provider as SanityProvider,
+  Query,
+  useSanity,
+  useSanityQuery
+} from "./sanity"
+import {
+  applyVariants,
+  borderRadius,
+  color,
+  commonPropTypes,
+  createGlobal,
+  createVariants,
+  cssResets,
+  flattenGlobals,
+  font,
+  fontFamily,
+  fontSmoothing,
+  fontWeight,
+  getProp,
+  gradient,
+  grid,
+  hideVisually,
+  logger,
+  mapProp,
+  spacing,
+  targetSize,
+  Theme,
+  theme,
+  themePropTypes,
+  themes,
+  VARIANT_DEFAULT,
+  wrapper
+} from "./theming"
+/**
+ * Utils
+ */
+import { distinct, flatten, partition, stringIncludes } from "./utils"
+/**
+ * i18n
+ */
+import {
+  languages,
+  Provider,
+  resourceReducer,
+  RESTBackend,
+  SanityBackend,
+  Transform,
+  Translate,
+  TranslateSSR,
+  useI18n
+} from "./i18n"
+/**
+ * Locale
+ */
+import { LocaleProvider, locales, useLocale } from "./locale"
+/**
+ * Hooks
+ */
+import {
+  useApiRequest,
+  useBrRegSearch,
+  useCopyToClipboard,
+  useDebounce,
+  useExchangeRates,
+  useInterval,
+  useLogging,
+  usePostalCode,
+  useSearch,
+  useTimer
+} from "./hooks"
+/**
+ * Components
+ */
+import {
+  Alert,
+  Anchor,
+  Box,
+  Breadcrumb,
+  Button,
+  Caret,
+  CaretRight,
+  Check,
+  CheckBox,
+  CheckGroup,
+  Close,
+  CodeRenderer,
+  Divider,
+  Donut,
+  Expand,
+  ExpandListItem,
+  FileInput,
+  Flag,
+  Fraction,
+  FullScreenLoader,
+  GlobalStyle,
+  Heading,
+  Image,
+  Info,
+  InlineListItem,
+  Input,
+  ItemGroup,
+  KeyFigures,
+  Label,
+  Layout,
+  LayoutItem,
+  Legend,
+  LinkListItem,
+  List,
+  Loading,
+  Logo,
+  Modal,
+  NewBadge,
+  Pack,
+  PackItem,
+  Paragraph,
+  ProgressBar,
+  QuestionMark,
+  RadioButton,
+  RadioPill,
+  RadioPillItem,
+  SelectSimple,
+  SkeletonCircle,
+  SkeletonLine,
+  SkeletonPulse,
+  Slider,
+  SliderKeyboardInput,
+  SliderTick,
+  SliderTrack,
+  SplitListItem,
+  Success,
+  Table,
+  Tag,
+  Text,
+  ThemeComponent,
+  ThemeProvider,
+  Toggle,
+  Warning,
+  WebFonts,
+  WidowFix,
+  Wrapper
+} from "./components"
+
+export const animations = {
+  AppearIn,
+  BounceIn,
+  BounceOut,
+  Draw,
+  FadeIn,
+  FadeInFromTop,
+  ScaleIn,
+  ScaleInX,
+  ScaleInY,
+  SlideInFromLeft,
+  SlideInFromRight,
+  SlideInFromTop
+}
+
+export const formatting = {
+  abbreviateCurrency,
+  CURRENCY_DISPLAY,
+  dashIt,
+  formatDate,
+  formatDateDistance,
+  formatDateRelative,
+  formatMoney,
+  formatName,
+  formatNumber,
+  maskFormat,
+  removeWhitespace,
+  STYLE,
+  truncate,
+  UNIT_DISPLAY,
+  UNITS
+}
+
+export const math = {
+  almostEquals,
+  clamp,
+  cosineInterpolate,
+  inverseLerp,
+  isNumber,
+  lerp,
+  multiplyPixelValue
+}
+
+export const openId = {
   useLogout,
   CallbackLogin,
   CallbackLoginSilent,
@@ -73,27 +279,24 @@ export {
   useOpenIdUser,
   useOtpLogin,
   useUrlLogin
-} from "./open-id"
+}
 
-/**
- * Sanity specific
- */
-export {
+export const sanity = {
   BlockContent,
   Document,
-  Image as SanityImage,
-  List as SanityList,
-  Provider as SanityProvider,
+  SanityImage,
+  SanityList,
+  SanityProvider,
   Query,
   useSanity,
   useSanityQuery
-} from "./sanity"
+}
 
 /**
  * Theming
  */
 
-export {
+export const theming = {
   VARIANT_DEFAULT,
   fontWeight,
   fontFamily,
@@ -120,16 +323,10 @@ export {
   themePropTypes,
   themes,
   wrapper
-} from "./theming"
-/**
- * Utils
- */
-export { distinct, flatten, partition, stringIncludes } from "./utils"
+}
+export const utils = { distinct, flatten, partition, stringIncludes }
 
-/**
- * i18n
- */
-export {
+export const i18n = {
   useI18n,
   languages,
   Provider,
@@ -139,17 +336,13 @@ export {
   Transform,
   Translate,
   TranslateSSR
-} from "./i18n"
+}
 
-/**
- * Locale
- */
-export { locales, LocaleProvider, useLocale } from "./locale"
+export {}
 
-/**
- * Hooks
- */
-export {
+export const locale = { locales, LocaleProvider, useLocale }
+
+export const hooks = {
   useApiRequest,
   useBrRegSearch,
   useCopyToClipboard,
@@ -160,12 +353,117 @@ export {
   usePostalCode,
   useSearch,
   useTimer
-} from "./hooks"
+}
 
-/**
- * Components
- */
 export {
+  AppearIn,
+  BounceIn,
+  BounceOut,
+  Draw,
+  FadeIn,
+  FadeInFromTop,
+  ScaleIn,
+  ScaleInX,
+  ScaleInY,
+  SlideInFromLeft,
+  SlideInFromRight,
+  SlideInFromTop,
+  abbreviateCurrency,
+  CURRENCY_DISPLAY,
+  dashIt,
+  formatDate,
+  formatDateDistance,
+  formatDateRelative,
+  formatMoney,
+  formatName,
+  formatNumber,
+  maskFormat,
+  removeWhitespace,
+  STYLE,
+  truncate,
+  UNIT_DISPLAY,
+  UNITS,
+  almostEquals,
+  clamp,
+  cosineInterpolate,
+  inverseLerp,
+  isNumber,
+  lerp,
+  multiplyPixelValue,
+  CallbackLogin,
+  CallbackLoginSilent,
+  CallbackLogout,
+  Logout,
+  OpenId,
+  OpenIdContext,
+  useCallbackLogin,
+  useCallbackLoginSilent,
+  useCallbackLogout,
+  useLogout,
+  useOpenId,
+  useOpenIdUser,
+  useOtpLogin,
+  useUrlLogin,
+  BlockContent,
+  Document,
+  SanityImage,
+  SanityList,
+  SanityProvider,
+  Query,
+  useSanity,
+  useSanityQuery,
+  applyVariants,
+  borderRadius,
+  color,
+  commonPropTypes,
+  createGlobal,
+  createVariants,
+  cssResets,
+  flattenGlobals,
+  font,
+  fontFamily,
+  fontSmoothing,
+  fontWeight,
+  getProp,
+  gradient,
+  grid,
+  hideVisually,
+  logger,
+  mapProp,
+  spacing,
+  targetSize,
+  Theme,
+  theme,
+  themePropTypes,
+  themes,
+  VARIANT_DEFAULT,
+  wrapper,
+  distinct,
+  flatten,
+  partition,
+  stringIncludes,
+  useI18n,
+  languages,
+  Provider,
+  resourceReducer,
+  RESTBackend,
+  SanityBackend,
+  Transform,
+  Translate,
+  TranslateSSR,
+  locales,
+  LocaleProvider,
+  useLocale,
+  useApiRequest,
+  useBrRegSearch,
+  useCopyToClipboard,
+  useDebounce,
+  useExchangeRates,
+  useInterval,
+  useLogging,
+  usePostalCode,
+  useSearch,
+  useTimer,
   Logo,
   Loading,
   Alert,
@@ -232,4 +530,4 @@ export {
   WebFonts,
   WidowFix,
   Wrapper
-} from "./components"
+}
